refactor: export FileInput
