+++ packages/client/src/shared/ui/index.ts
@@ -1,1 +1,2 @@
 export { CounterAnimation } from './CounterAnimation'
+export { FileInput } from './FileInput'
